Module migration_connector::migrations_directory

Migrations directory management.

This module is responsible for the management of the contents of the migrations directory. At the top level it contains a migration_lock.toml file which lists the provider. It also contains multiple subfolders, named after the migration id, and each containing:

A migration script

Structs

ListMigrationsError

An IO error that occurred while reading the migrations directory.

MigrationDirectory

Proxy to a directory containing one migration, as returned by create_migration_directory and list_migrations.

ReadMigrationScriptError

Error while reading a migration script.

Constants

MIGRATION_LOCK_FILENAME

The file name for the migration lock file, not including the file extension.

MIGRATION_SCRIPT_FILENAME

The file name for migration scripts, not including the file extension.

Functions

create_migration_directory

Create a directory for a new migration.

error_on_changed_provider

Error if the provider in the schema does not match the one in the schema_lock.toml

list_migrations

List the migrations present in the migration directory, lexicographically sorted by name.

write_migration_lock_file

Write the migration_lock file to the directory.