Trait migration_connector::MigrationConnector

pub trait MigrationConnector: Send + Sync + 'static {

    fn set_host(&mut self, host: Arc<dyn ConnectorHost>);

    fn acquire_lock<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn connector_type(&self) -> &'static str;

    fn connection_string(&self) -> &str;

    fn create_database<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<String>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn db_execute<'life0, 'async_trait>(
        &'life0 self, 
        url: String, 
        script: String
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn diff<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 self, 
        from: DiffTarget<'life1>, 
        to: DiffTarget<'life2>
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<Migration>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    fn drop_database<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn ensure_connection_validity<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn host(&self) -> &Arc<dyn ConnectorHost>;

    fn version<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<String>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn reset<'life0, 'async_trait>(
        &'life0 self
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn migration_file_extension(&self) -> &'static str;

    fn migration_len(&self, migration: &Migration) -> usize;

    fn migration_persistence(&self) -> &dyn MigrationPersistence;

    fn migration_summary(&self, migration: &Migration) -> String;

    fn database_migration_step_applier(
        &self
    ) -> &dyn DatabaseMigrationStepApplier;

    fn destructive_change_checker(&self) -> &dyn DestructiveChangeChecker;

    fn validate_migrations<'life0, 'life1, 'async_trait>(
        &'life0 self, 
        _migrations: &'life1 [MigrationDirectory]
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn check_database_version_compatibility(
        &self, 
        _datamodel: &ValidatedSchema
    ) -> Option<DatabaseVersionIncompatibility> { ... }

    fn migration_is_empty(&self, migration: &Migration) -> bool { ... }

}

The top-level trait for connectors. This is the abstraction the migration engine core relies on to interface with different database backends.

Required methods

fn set_host(&mut self, host: Arc<dyn ConnectorHost>)

Accept a new ConnectorHost.

fn acquire_lock<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

If possible on the target connector, acquire an advisory lock, so multiple instances of migrate do not run concurrently.

fn connector_type(&self) -> &'static str

A string that should identify what database backend is being used. Note that this is not necessarily the connector name. The SQL connector for example can return “postgresql”, “mysql” or “sqlite”.

fn connection_string(&self) -> &str

Return the connection string that was used to initialize this connector.

fn create_database<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = ConnectorResult<String>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Create the database referenced by Prisma schema that was used to initialize the connector.

fn db_execute<'life0, 'async_trait>(
&'life0 self,
url: String,
script: String
) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Send a command to the database directly.

fn diff<'life0, 'life1, 'life2, 'async_trait>(
&'life0 self,
from: DiffTarget<'life1>,
to: DiffTarget<'life2>
) -> Pin<Box<dyn Future<Output = ConnectorResult<Migration>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Create a migration by comparing two database schemas. See DiffTarget for possible inputs.

fn drop_database<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Drop the database referenced by Prisma schema that was used to initialize the connector.

fn ensure_connection_validity<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Make sure the connection to the database is established and valid. Connectors can choose to connect lazily, but this method should force them to connect.

fn host(&self) -> &Arc<dyn ConnectorHost>

Return the ConnectorHost passed with set_host.

fn version<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = ConnectorResult<String>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

The version of the underlying database.

fn reset<'life0, 'async_trait>(
&'life0 self
) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Drop all database state.

fn migration_file_extension(&self) -> &'static str

The file extension for generated migration files.

fn migration_len(&self, migration: &Migration) -> usize

Return the number of steps in the migration. Invariant: migration_is_empty() == true iff migration_len() == 0.

fn migration_persistence(&self) -> &dyn MigrationPersistence

See MigrationPersistence.

fn migration_summary(&self, migration: &Migration) -> String

Render a human-readable drift summary for the migration.

fn database_migration_step_applier(&self) -> &dyn DatabaseMigrationStepApplier

See DatabaseMigrationStepApplier.

fn destructive_change_checker(&self) -> &dyn DestructiveChangeChecker

See DestructiveChangeChecker.

fn validate_migrations<'life0, 'life1, 'async_trait>(
&'life0 self,
_migrations: &'life1 [MigrationDirectory]
) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

If possible, check that the passed in migrations apply cleanly.

Provided methods

fn check_database_version_compatibility(
&self,
_datamodel: &ValidatedSchema
) -> Option<DatabaseVersionIncompatibility>

Optionally check that the features implied by the provided datamodel are all compatible with the specific database version being used.

fn migration_is_empty(&self, migration: &Migration) -> bool

Return whether the migration is empty.