Trait migration_connector::MigrationPersistence

pub trait MigrationPersistence: Send + Sync {
    fn baseline_initialize<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn initialize<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn mark_migration_applied_impl<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self, 
        migration_name: &'life1 str, 
        checksum: &'life2 str
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<String>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    fn mark_migration_rolled_back_by_id<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        migration_id: &'life1 str
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn record_migration_started_impl<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self, 
        migration_name: &'life1 str, 
        checksum: &'life2 str
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<String>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    fn record_successful_step<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        id: &'life1 str
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn record_failed_step<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self, 
        id: &'life1 str, 
        logs: &'life2 str
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;

    fn record_migration_finished<'life0, 'life1, 'async_trait>(
        &'life0 mut self, 
        id: &'life1 str
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn list_migrations<'life0, 'async_trait>(
        &'life0 mut self
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<Result<Vec<MigrationRecord>, PersistenceNotInitializedError>>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        Self: 'async_trait;

    fn mark_migration_applied<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self, 
        migration_name: &'life1 str, 
        script: &'life2 str
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<String>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait,
    { ... }

    fn record_migration_started<'life0, 'life1, 'life2, 'async_trait>(
        &'life0 mut self, 
        migration_name: &'life1 str, 
        script: &'life2 str
    ) -> Pin<Box<dyn Future<Output = ConnectorResult<String>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait,
    { ... }
}

Management of imperative migrations state in the database.

Required methods

fn baseline_initialize<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

Initialize the migration persistence without checking the database first.

fn initialize<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

This method is responsible for checking whether the migrations persistence is initialized.

If the migration persistence is not present in the target database, check whether the database schema is empty. If it is, initialize the migration persistence. If not, return a DatabaseSchemaNotEmpty error.

fn mark_migration_applied_impl<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
migration_name: &'life1 str,
checksum: &'life2 str
) -> Pin<Box<dyn Future<Output = ConnectorResult<String>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Implementation in the connector for the core’s MarkMigrationApplied command. See the docs there. Note that the started_at and finished_at for the migration should be the same.

fn mark_migration_rolled_back_by_id<'life0, 'life1, 'async_trait>(
&'life0 mut self,
migration_id: &'life1 str
) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Mark the failed instances of the migration in the persistence as rolled back, so they will be ignored by the engine in the future.

fn record_migration_started_impl<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
migration_name: &'life1 str,
checksum: &'life2 str
) -> Pin<Box<dyn Future<Output = ConnectorResult<String>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Record that a migration is about to be applied. Returns the unique identifier for the migration.

This is an implementation detail, consumers should use record_migration_started() instead.

fn record_successful_step<'life0, 'life1, 'async_trait>(
&'life0 mut self,
id: &'life1 str
) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Increase the applied_steps_count counter.

fn record_failed_step<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
id: &'life1 str,
logs: &'life2 str
) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Report logs for a failed migration step. We assume the next steps in the migration will not be applied, and the error reported.

fn record_migration_finished<'life0, 'life1, 'async_trait>(
&'life0 mut self,
id: &'life1 str
) -> Pin<Box<dyn Future<Output = ConnectorResult<()>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Record that the migration completed successfully. This means populating the finished_at field in the migration record.

fn list_migrations<'life0, 'async_trait>(
&'life0 mut self
) -> Pin<Box<dyn Future<Output = ConnectorResult<Result<Vec<MigrationRecord>, PersistenceNotInitializedError>>> + Send + 'async_trait>> where
'life0: 'async_trait,
Self: 'async_trait,

List all applied migrations, ordered by started_at. This should fail with a PersistenceNotInitializedError when the migration persistence is not initialized.

Provided methods

fn mark_migration_applied<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
migration_name: &'life1 str,
script: &'life2 str
) -> Pin<Box<dyn Future<Output = ConnectorResult<String>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Implementation in the connector for the core’s MarkMigrationApplied command. See the docs there. Note that the started_at and finished_at for the migration should be the same.

Connectors should implement mark_migration_applied_impl to avoid doing the checksuming themselves.

fn record_migration_started<'life0, 'life1, 'life2, 'async_trait>(
&'life0 mut self,
migration_name: &'life1 str,
script: &'life2 str
) -> Pin<Box<dyn Future<Output = ConnectorResult<String>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Record that a migration is about to be applied. Returns the unique identifier for the migration.

This is a default method that computes the checksum. Implementors should implement record_migration_started_impl.