Crate migration_connector

This crate defines the API exposed by the connectors to the migration engine core. The entry point for this API is the MigrationConnector trait.

Modules

migrations_directory

Migrations directory management.

Structs

ConnectorError

The general error reporting type for migration connectors.

ConnectorParams

Parameters passed from the core to connectors on initialization.

DatabaseSchema

A database schema. Part of the MigrationConnector API.

DestructiveChangeDiagnostics

The errors and warnings emitted by the DestructiveChangeChecker.

EmptyHost

A no-op ConnectorHost.

Migration

A boxed migration, opaque to the migration engine core. The connectors are sole responsible for producing and understanding migrations — the core just orchestrates.

MigrationRecord

An applied migration, as returned by list_migrations.

MigrationWarning

A warning emitted by DestructiveChangeChecker. Warnings will prevent a migration from being applied, unless the force flag is passed.

PersistenceNotInitializedError

Error returned when the persistence is not initialized.

UnexecutableMigration

An unexecutable migration step detected by the DestructiveChangeChecker.

Enums

DiffTarget

Diffable things

Traits

ConnectorHost

An abstract host for a migration connector. It exposes IO that is not directly performed by the connectors.

DestructiveChangeChecker

Implementors of this trait are responsible for checking whether a migration could lead to data loss, or if it would be potentially unexecutable.

MigrationConnector

The top-level trait for connectors. This is the abstraction the migration engine core relies on to interface with different database backends.

MigrationPersistence

Management of imperative migrations state in the database.

Type Definitions

BoxFuture

Alias for a pinned, boxed future, used by the traits.

ConnectorResult

Shorthand for a Result where the error variant is a ConnectorError.

Timestamp

A timestamp.